backend/requirements.txt
Django==5.2.1
psycopg2-binary==2.9.9
djangorestframework==3.15.1  
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
gunicorn==21.2.0            
python-dotenv==1.0.1        
asgiref==3.8.1              
sqlparse==0.5.3    
Pillow>=10.0.0